+++ src/main/java/de/maxhenkel/easyvillagers/integration/IMC.java
@@ -1,6 +1,6 @@
 package de.maxhenkel.easyvillagers.integration;
 
-import de.maxhenkel.easyvillagers.integration.theoneprobe.TheOneProbeModule;
+//import de.maxhenkel.easyvillagers.integration.theoneprobe.TheOneProbeModule;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.InterModComms;
 import net.minecraftforge.fml.ModList;
@@ -10,9 +10,9 @@ public class IMC {
 
     @SubscribeEvent
     public static void enqueueIMC(InterModEnqueueEvent event) {
-        if (ModList.get().isLoaded("theoneprobe")) {
+        /*if (ModList.get().isLoaded("theoneprobe")) {
             InterModComms.sendTo("theoneprobe", "getTheOneProbe", TheOneProbeModule::new);
-        }
+        }*/
     }
 
 }